refactor: Consolidate CRU tests and remove redundant CSV file
- Merge test_cru_basic.py and test_cru_data_formats.py into single test_cru.py
- Remove 19MB CSV file from test fixtures (already verified equivalence with Parquet)
- Keep only essential tests: data loading, temperature lookup, and precheck integration
- Reduced from 2 test files to 1, maintaining all 6 essential tests
- Saves 19MB from repository size by removing redundant CSV
